requirements.txt
Flask==3.0.3
requests==2.32.3
openai==1.41.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.104
pinecone-client==5.0.1
pinecone-notebooks==0.1.1